services/rest_api/requirements.txt
alembic==0.8.5
aniso8601==1.1.0
blinker==1.4
boto3==1.4.0
botocore==1.4.56
click==6.4
filemagic==1.6
Flask==0.10.1
Flask-Cache==0.13.1
Flask-Login==0.3.2
Flask-Migrate==1.8.0
Flask-OAuthlib==0.9.2
Flask-RESTful==0.3.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
future==0.15.2
gunicorn==19.4.5
inflection==0.3.1
itsdangerous==0.24
Jinja2==2.8
Mako==1.0.4
MarkupSafe==0.23
marshmallow==2.6.1
nose==1.3.7
oauthlib==1.0.3
pip==9.0.1
psycopg2==2.6.1
PyGithub==1.29
python-dateutil==2.5.2
python-editor==0.5
pytz==2016.3
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.1
rq==0.7.0
six==1.10.0
SQLAlchemy==1.0.12
Werkzeug==0.11.5
WTForms==2.1
rsyslog
